fix: Change w,h in BoundingBox f32 as well

+++ src/classes.rs
@@ -2,10 +2,10 @@
 use fake::{Dummy, Fake};
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Default, Clone, Serialize, Deserialize, Dummy, PartialEq, Eq)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize, Dummy, PartialEq)]
 pub struct BoundingBox {
     pub x: i32,
     pub y: i32,
-    pub w: u32,
-    pub h: u32,
+    pub w: f32,
+    pub h: f32,
 }
